requirements.txt
Flask==3.0.3
Flask-MySQLdb==2.0.0
Jinja2==3.1.4
mysqlclient==2.2.6
Werkzeug==3.1.3
python-dotenv==1.0.1
Flask==3.0.3
Flask-WTF==1.2.1
WTForms==3.1.2
